spi: move CloverImage
